Backyard Garden
Backyard Garden will be an educational android game application that
targets young children before and during kindergarten. In this game
children will be able to interact with a backyard garden setting and
have fun growing plants while learning basic math and science. There
will be no “ending” to our game but it will have set up achievements to
complete. After completing certain achievements, the game will introduce
new educational concepts so that the child playing will be able to keep
learning throughout the game at their own pace. We believe that this
method will allow a child to desire learning rather than feel
discouraged about what they are doing.

The game will begin with a friendly backyard setting with a plot of land
designated for planting. The user will be prompted to place one seed in
the dirt, water it, and give it some sunshine. They will then be able to
see the seed grow into a seedling, and then the user will repeat the
water and sunshine process. After watering the plant and giving it
sunshine twice, the plant will bloom into a flower and the user will
have a basic understanding of plant growth. The user will be able to
count how many seeds they need with a visual representation. This will
begin teaching the user how to count from 1 to 10.

Currently, there is no replica game of our project idea and we are
excited to offer it to parents. We believe that they will appreciate and
support our aim of teaching basic math and science to children in a fun
new way. Children are becoming familiar with trending technology at a
higher rate and we believe that utilizing the android market to showcase
our project will help us gain popularity and profit.

For this semester, we will focus on the most basic concepts as described
above. However, we envision new educational topics added to a series of
this game application in the future. This will expand our consumer base
and allow our current consumers to grow with us. The success of these
applications could eventually lead our team to try selling our product
to private and public schools and will be great to present to
homeschooling families.